+++ division.c
@@ -223,11 +223,11 @@ enum eExprErr BigIntDivide(const BigInteger *pDividend, const BigInteger *pDivis
       }
       if (low < HALF_INT_RANGE)
       {
-        carry = (int)floor((dAccumulator + HALF_INT_RANGE / 2) * dVal);
+        carry = (int)floor((dAccumulator + (double)(HALF_INT_RANGE / 2)) * dVal);
       }
       else
       {
-        carry = (int)floor((dAccumulator - HALF_INT_RANGE / 2) * dVal);
+        carry = (int)floor((dAccumulator - (double)(HALF_INT_RANGE / 2)) * dVal);
       }
       ptrDividend->x = low;
 #endif
